% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results_table.R
\name{extract_result_table}
\alias{extract_result_table}
\title{Extract the results table from GMRFILM output}
\usage{
extract_result_table(sim_dir, sim_fi, skip_top, n_rows, verbose = FALSE)
}
\arguments{
\item{sim_dir}{A string. The path to the simulation directory.
Example: here()}

\item{sim_fi}{A string. The base file name of the GMRFILM output
that was piped to a file. Example: "case-50a". The
function appends '.out'}

\item{skip_top}{An integer. The number of rows when reading in the
raw output from the .out file. Example: 55. You want
to capture the title line for readability.}

\item{n_rows}{An integer. The number of rows (lines) to keep.
Example: 9. This should yield a complete output
table.}

\item{verbose}{A Boolean. Default: FALSE. If TRUE it prints lines.}
}
\value{
None.   The function writes the extracted table to 
                the sim_fi variable with '-ext.txt' appended
}
\description{
Extract the results table from GMRFILM output
}
\examples{
library(rgmrfilm)
# to be added
}
