% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_empty_values.R
\name{no_empty_values}
\alias{no_empty_values}
\title{Remove empty values from a strsplt() call for easier processing}
\usage{
no_empty_values(li)
}
\arguments{
\item{li}{A list of strings output from strsplt() containing missing
values (i.e. "")}
}
\value{
li The list without the missing values
}
\description{
Remove empty values from a strsplt() call for easier processing
}
\examples{
library(rgmrfilm)
# to be added
}
