% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_multiple_spaces.R
\name{remove_multiple_spaces}
\alias{remove_multiple_spaces}
\title{Remove mutiple spaces from a string}
\usage{
remove_multiple_spaces(str)
}
\arguments{
\item{str}{A string. The string to be cleaned}
}
\value{
str  The cleaned string
}
\description{
Remove mutiple spaces from a string
}
\examples{
# library(rgmrfilm)
#
}
